Move scripts/download to scripts/dependencies
